Refactor test assets & fixtures
[ci skip]
