+++ f/frizzle/kobo/tests/assets/trees.png
