+++ f/frizzle/kobo/tests/kobo_responses_test.py
@@ -1,10 +1,4 @@
-import re
-from dataclasses import dataclass
-
-import pytest
-
 from f.frizzle.kobo.kobo_responses import main, sanitize
-from f.frizzle.kobo.tests import kobo_form
 
 
 def test_sanitize():
@@ -85,44 +79,6 @@ def test_sanitize_with_nesting():
         "group1": '{"group2": {"question": "How ya doin?"}}',
         "url": "gopher://example.net",
     }
-
-
-@pytest.fixture
-def koboserver(mocked_responses):
-    """A mock Kobo Server that you can use to provide survey responses"""
-
-    @dataclass
-    class KoboServer:
-        account: dict
-        form_id: str
-
-    server_url = "http://kobo.example.org"
-    form_id = "mimsyweretheborogoves"
-    form_name = "Arboles"
-
-    mocked_responses.get(
-        f"{server_url}/api/v2/assets/{form_id}/",
-        json=kobo_form.kobo_form(server_url, form_id, form_name),
-        status=200,
-    )
-    mocked_responses.get(
-        f"{server_url}/api/v2/assets/{form_id}/data/",
-        json=kobo_form.kobo_form_submissions(server_url, form_id),
-        status=200,
-    )
-    mocked_responses.get(
-        re.compile(rf"{server_url}/api/v2/assets/{form_id}/data/\d+/attachments/\d+/?"),
-        body=open("f/frizzle/kobo/tests/trees.png", "rb").read(),
-        content_type="image/png",
-        headers={"Content-Length": "3632"},
-    )
-
-    return KoboServer(dict(server_url=server_url, api_key="Callooh!Callay!"), form_id)
-
-
-@pytest.fixture
-def pg_database():
-    return dict(dbname="", user="", password="", host="", port="")
 
 
 def test_script(koboserver, pg_database, tmp_path):
